requirements.txt
django~=2.2
psycopg2==2.8.6
Telethon==1.17.5
djangorestframework==3.12.1